Fixes edge cases when using the binary serializer
### Polymorphic handlers

Sending a type that is serialized with the binary serializer and conumed by its base type
on another bus instance configured with `SupportBinaryMessageDeserializer` and
another serializer

### Serializer negotiation on faults

If a fault is produced during while processing the message the generated fault is serialized
using the current serializer configured on the bus.
This should be configurable since the serialization can fail with a different serializer. Using the same serializer used to
deserialize the message guarantee that the serialization succeed.

**Note** that to accomplish this we need to change the fault generation process. With the current extensions points we
need to:
    * Decorate the ConsumeContext
    * Forward all calls except
        * `ConsumeContext.NotifyFault<T>`: Here we only need to override the serializer based on the content type in the
            receive context
         * `ConsumeContext.TryGetMessage<T>` to ensure that when a ConsumeContext become typed we can still
             intercept `ConsumeContext.NotifyFault<T>`

All of this is just boiler plate code (we cannot use `ConsumeContextProxy` because `CompleteTask` it is not virtual and
we need to override it to reimplement the `ConsumeContext.NotifyFault<T>` behavior).
What we really need is an extension point in the default `GenerateFault` code that pass the `ReceiveContext` and the `SendContext<Fault<T>>`  that is being constructed (see `SerializerNegotiatorForFaultConsumeContextDecorator.PrepareFaultSendContext`).

Ideally it would be nice to be able to decorate the default implementation and add this behavior during the
configuration of the pipe used to send the fault.

Another alternative would be to inject the ReceiveContext(or ConsumeContext) as payload when creating the
`SendContext` is created.
This would allow the same implementation using filters over the send and publish endpoints by extracting the source
context from the input `SendContext` instance that is passed to the filter and performing the negotiation there.
This has the disavantange that the filter would be executed for every send and publish operation instead of just for faults.
Obviously the serializer would be overriden only If `TryGetPayload<ReceiveContext>` return true, but still.